Move instantiate_opaque_types to rustc_infer.
It does not depend on anything from rustc_trait_selection anymore.
